requirements.txt
asgiref==3.8.1
astroid==2.4.2
backports.zoneinfo==0.2.1
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
Faker==26.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
python-dateutil==2.9.0.post0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing_extensions==4.12.2
validate-docbr==1.10.0
wrapt==1.12.1
